// generated/src/app/api-models/auth-type-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthTypeDetail {
  seqAuthType: number;
  authScreen: string;
  screenRank: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}